// src/api/generated/schemas/chatSessionMessage.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * The Bhakti Vault API
 * FastAPI application with OpenID Connect authentication
 * OpenAPI spec version: 1.0.0
 */
import type { ChatRole } from './chatRole';
import type { ChatObject } from './chatObject';

export interface ChatSessionMessage {
  public_id: string;
  chat_session_id: string;
  role: ChatRole;
  content: ChatObject[];
  created_on: number;
}